% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_coherence}
\alias{plot_coherence}
\title{Plotting Coherence Measure}
\usage{
plot_coherence(results)
}
\arguments{
\item{results}{The Data Frame containing the CV results}
}
\value{
A plot showing the mean and lower/upper bounds for each topic number included.
}
\description{
Plot coherence results. For best results, do not change column names.
}
\examples{
plot <- plot_coherence(results)
}
