% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tenFoldCV}
\alias{tenFoldCV}
\title{10-Fold Cross-Validation for Topic Modeling}
\usage{
tenFoldCV(dtm, range_of_k, burnin_value, iter_value, keep_value, set_seed)
}
\arguments{
\item{dtm}{The document term matrix.}

\item{range_of_k}{Range of topic numbers to test using LDA model.}

\item{burnin_value}{The number of iterations to discard at the beginning of the MCMC sampling process.}

\item{iter_value}{The total number of iterations for the MCMC sampling process.}

\item{keep_value}{The thinning parameter for the MCMC sampling process.}

\item{set_seed}{The seed used to ensure replicability.}
}
\value{
A Data Frame with the mean and confidence intervals for all three metrics: perplexity, likelihood, and semantic coherence.
}
\description{
Run a 10-fold CV for topic modeling using the `topicmodels` package. The method used is Gibbs sampling and the function will return the perplexity, log-likelihood, and topic coherence for each value of k (number of topics) in the candidate_k vector. The function will return the mean and the confidence intervals for each measure in a Data Frame.
}
\examples{
results <- tenFoldCV(dtm, seq(3,30,1), 500, 1000, 100, 444)
}
